more json web token strategy

+++ models/catModel.js
@@ -1,68 +1,73 @@
-'use strict';
-const pool = require('../database/db');
-const { httpError } = require('../utils/errors');
+"use strict";
+const pool = require("../database/db");
+const { httpError } = require("../utils/errors");
 const promisePool = pool.promise();
 
 const getAllCats = async (next) => {
   try {
     // TODO: do the LEFT (or INNER) JOIN to get owner's name as ownername (from wop_user table).
-    const [rows] = await promisePool.execute('SELECT * FROM wop_cat');
+    const [rows] =
+      await promisePool.execute(`SELECT cat_id, wop_cat.name, weight, wop_cat.owner, filename, birthdate, wop_user.name AS ownername FROM wop_cat 
+    INNER JOIN wop_user
+    ON wop_cat.owner = wop_user.user_id;`);
     return rows;
   } catch (e) {
-    console.error('getAllCats error', e.message);
-    next(httpError('Database error', 500));
+    console.error("getAllCats error", e.message);
+    next(httpError("Database error", 500));
   }
 };
 
 const getCat = async (id, next) => {
   try {
     const [rows] = await promisePool.execute(
-      'SELECT * FROM wop_cat WHERE cat_id = ?',
+      `SELECT cat_id, wop_cat.name, weight, wop_cat.owner, filename, birthdate, wop_user.name AS ownername FROM wop_cat 
+    INNER JOIN wop_user
+    ON wop_cat.owner = wop_user.user_id WHERE cat_id = ?;`,
       [id]
     );
     return rows;
   } catch (e) {
-    console.error('getCat error', e.message);
-    next(httpError('Database error', 500));
+    console.error("getCat error", e.message);
+    next(httpError("Database error", 500));
   }
 };
 
 const addCat = async (name, weight, owner, birthdate, filename, next) => {
   try {
     const [rows] = await promisePool.execute(
-      'INSERT INTO wop_cat (name, weight, owner, filename, birthdate) VALUES (?, ?, ?, ?, ?)',
+      "INSERT INTO wop_cat (name, weight, owner, filename, birthdate) VALUES (?, ?, ?, ?, ?)",
       [name, weight, owner, filename, birthdate]
     );
     return rows;
   } catch (e) {
-    console.error('addCat error', e.message);
-    next(httpError('Database error', 500));
+    console.error("addCat error", e.message);
+    next(httpError("Database error", 500));
   }
 };
 
 const modifyCat = async (name, weight, owner, birthdate, cat_id, next) => {
   try {
     const [rows] = await promisePool.execute(
-      'UPDATE wop_cat SET name = ?, weight = ?, owner = ?, birthdate = ? WHERE cat_id = ?;',
-      [name, weight, owner, birthdate, cat_id]
+      "UPDATE wop_cat SET name = ?, weight = ?, birthdate = ? WHERE cat_id = ? AND owner = ?;",
+      [name, weight, birthdate, cat_id, owner]
     );
     return rows;
   } catch (e) {
-    console.error('addCat error', e.message);
-    next(httpError('Database error', 500));
+    console.error("addCat error", e.message);
+    next(httpError("Database error", 500));
   }
 };
 
-const deleteCat = async (id, next) => {
+const deleteCat = async (id, owner_id, next) => {
   try {
     const [rows] = await promisePool.execute(
-      'DELETE FROM wop_cat WHERE cat_id = ?',
-      [id]
+      "DELETE FROM wop_cat WHERE cat_id = ? AND owner = ?",
+      [id, owner_id]
     );
     return rows;
   } catch (e) {
-    console.error('getCat error', e.message);
-    next(httpError('Database error', 500));
+    console.error("getCat error", e.message);
+    next(httpError("Database error", 500));
   }
 };
